requirements.txt
crewai
crewai-tools
langchain
openai
python-dotenv

fastapi
uvicorn

pydantic
redis
sqlalchemy

pytest
httpx

loguru
rich
litellm